==> main/01 - raw_data.ipynb <==
# BigQuery Schema

import os
import pandas as pd
from pathlib import Path
from google.cloud import bigquery

from targetandmarket.config import data_folder

# BigQuery client

SERVICE_ACCOUNT = os.environ['GOOGLE_API_KEY']
bqclient = bigquery.Client.from_service_account_json(SERVICE_ACCOUNT)

**Preliminary look at raw data stored on BigQuery**

# For user behaviour I am interested in the user_engagement events.
# Query from events of May 01, 2020
query = """
SELECT user_id, event_timestamp, event_name, event_params
FROM `analytics_157832975.events_20200501`
WHERE event_name = 'user_engagement'
LIMIT 100
"""

beh_data = pd.DataFrame(bqclient.query(query).result().to_dataframe())

beh_data.head()

**For every user_engagement event, event_params are recorded which are a list of dictionaries. Event_params contains the session ID (ga_session_id) and engagement time (engagement_time_msec) which are of interest to me.**

beh_data.iloc[0, 3]

**Next step:**

Total data size is 125 GB with over 36 million "user_engagement" events. The next step is to query engineered features that could have some predictive power.
