==> main/03 - clean_merge_data.ipynb <==
# Clean and merge data

**This section cleans and merges activity data from BigQuery and investment portfolio data from MySQL into one data frame.**

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from targetandmarket.config import data_folder

# Import app activity data
# Note that session numbers and engagement times have been scaled by a hidden factor to obfuscate true values
user_analytics_raw = pd.read_csv(data_folder/'user_analytics_raw.csv', index_col=0)
user_error = pd.read_csv(data_folder/'user_error_status.csv', index_col=0)
timeseries = pd.read_csv(data_folder/'timeseries.csv', index_col=0)

**Let's look at the activity data which I queried from BigQuery**

print(user_analytics_raw.shape)

user_analytics_raw.head()

**I have queried:**
* user_id: obfuscated user ID
* sessions: total number of sessions
* last_session: unix timestamp of last session (seconds)
* total_time: total engagement time
* avg_session: average session time
* first_open: unix timestamp of when app was first opened

**Session numbers were absent for some users so any user which such a session is removed**

user_analytics = user_analytics_raw.loc[~user_analytics_raw['user_id'].
                                        isin(user_error['user_id'])].copy()
print(user_analytics.shape)

**Since I am interested in predicting which users are paying users based on activity from May 2020, I need to exclude those users that started using the app in the middle of May 2020**

# Users who started using the app since app launch and 01 May 2020 are selected
PERIOD_START = pd.Timestamp('2017-11-01').timestamp()
PERIOD_END = pd.Timestamp('2020-05-01').timestamp()
user_analytics = user_analytics.loc[
    (user_analytics['first_open'] >= PERIOD_START)
    & (user_analytics['first_open'] < PERIOD_END)].copy()
print('Number of users between analysis duration:', len(user_analytics))

**Converting unix timestamps to "delta" between June 01, 2020 (in terms of number of days ago)**

LAST_MODIFIED = 1590969600
user_analytics.loc[:, 'last_session'] = (
    (LAST_MODIFIED - user_analytics['last_session']) / (60 * 60 * 24))
user_analytics.loc[:, 'first_open'] = (
    (LAST_MODIFIED - user_analytics['first_open']) / (60 * 60 * 24))
user_analytics = user_analytics.astype('float64')

**Calculating the number of days a user was active in May 2020 from timeseries data**

timeseries = timeseries.pivot(index='user_id',
                              columns='date',
                              values='day_eng_time')
active_days = pd.notna(timeseries).sum(axis=1)
active_days.name = 'active_days'

user_analytics = pd.merge(user_analytics,
                          active_days,
                          how='inner',
                          on='user_id')

user_analytics.head()

**Now let's look at the investment data and our label (free vs paying user)**

user_holdings = pd.read_csv(data_folder/'user_holdings.csv')
user_purchases = pd.read_csv(data_folder/'user_purchases_may.csv', index_col=0)
user_purchases_june = pd.read_csv(data_folder/'user_purchases_early_june.csv', index_col=0)
user_transactions = pd.read_csv(data_folder/'user_transactions.tsv', sep='\t')

user_transactions.head()

#%% Merge number of transactions and number of coins with user analytics
user_analytics = pd.merge(user_analytics,
                          user_transactions[['obfuscatedId', 'numberOfTransactions', 'numberOfCoins']],
                          how='inner',
                          left_on='user_id',
                          right_on='obfuscatedId')
user_analytics = user_analytics.drop(['obfuscatedId'], axis=1)

user_holdings.head()

user_holdings.describe()

**There are some outliers when it comes to holdings (value of investment portfolio). There are some users with negative holdings and some with holdings greater than the market capitalization of all digital assets. So I will only keep users with more than 10 USD value and less than 10 million USD**

user_holdings = user_holdings.loc[(user_holdings.holdings < 10000000)
                                  & (user_holdings.holdings > 10)]

# Merge with activity data
user_analytics = pd.merge(user_analytics,
                          user_holdings[['obfuscatedId', 'holdings', 'isPro']],
                          how='inner',
                          left_on='user_id',
                          right_on='obfuscatedId')
user_analytics = user_analytics.drop(['obfuscatedId'], axis=1)

**The isPro column indicates which user is currently a "PRO" or premium user. It also includes users who are on free trial (7 days) so these are actually free users and I need to change the label to 0**

# Purchase history from May 2020
user_purchases = user_purchases.dropna()
user_purchases = user_purchases.drop(['event_name'], axis=1)

mapping = {
    'io.getdelta.android.delta_pro_yearly_trial': 0,
    'io.getdelta.ios.DELTA_PRO_EARLY_BACKER_MONTHLY': 1,
    'io.getdelta.ios.DELTA_PRO_YEARLY_TRIAL': 0,
    'io.getdelta.ios.DELTA_PRO_EARLY_BACKER_MONTHLY_EQUALIZED': 1,
    'io.getdelta.ios.DELTA_PRO_EARLY_BACKER_YEARLY': 1,
    'io.getdelta.android.delta_pro_early_backer_yearly_equalized': 1
}

user_purchases.loc[:, 'product_id'] = user_purchases['product_id'].map(mapping)
user_purchases = user_purchases.sort_values(by=['user_id', 'product_id'])
# Keep last product "purchased" because those that actually purchased premium after trial should be left as is
user_purchases = user_purchases.drop_duplicates(subset=['user_id'],
                                                keep='last')
user_analytics = pd.merge(user_analytics,
                          user_purchases,
                          how='left',
                          on='user_id')
# Users that have pro status = 1 but are on trial (product_id ==0), set pro status as 0
user_analytics.loc[((user_analytics.isPro == 1) &
                    (user_analytics.product_id == 0)), 'isPro'] = 0
user_analytics = user_analytics.drop(['product_id'], axis=1)

**The isPro status is as of June 03, 2020. Due to the way the company stores data, historical status of isPro is not available. Therefore, I have used purchase history on June 01, 02 and 03 to set users that purchased premium during these days as non-premium for May 2020**

user_purchases_june = user_purchases_june.drop(['event_name'], axis=1)
user_purchases_june = user_purchases_june.dropna()
user_purchases_june.loc[:, 'product_id'] = user_purchases_june['product_id'].map(mapping)
user_purchases_june = user_purchases_june.sort_values(by=['user_id', 'product_id'])
user_purchases_june = user_purchases_june.drop_duplicates(subset=['user_id'], keep='last')
user_analytics = pd.merge(user_analytics, user_purchases_june, how='left', on='user_id')
# Set pro status for those who purchased or tried after June 01 as 0
user_analytics.loc[((user_analytics.isPro==1) & (~pd.isna(user_analytics.product_id))), 'isPro'] = 0
user_analytics = user_analytics.drop(['product_id'], axis=1)

user_analytics.head()

#%% Save processed data
user_analytics.to_csv(data_folder/'user_analytics.csv')
